feat: Introduce EmailNotifierMethods concern to share email methods across notifiers
