+++ app/notifiers/base_notifier.rb
@@ -1,11 +1,8 @@
 class BaseNotifier < AbstractNotifier
+  include EmailNotifierMethods
+
   def created
-    EnrollmentMailer.with(
-      to: enrollment.user.email,
-      target_api: enrollment.target_api,
-      enrollment_id: enrollment.id,
-      template: "create_application"
-    ).notification_email.deliver_later
+    deliver_created_mail_to_enrollment_creator
   end
 
   def updated(diff:, user_id:)
@@ -15,25 +12,16 @@ class BaseNotifier < AbstractNotifier
   end
 
   def rgpd_contact_updated(diff:, user_id:, responsable_traitement_email:, dpo_email:)
-    if responsable_traitement_email
-      deliver_rgpd_email_for(:responsable_traitement)
-    end
-
-    if dpo_email
-      deliver_rgpd_email_for(:dpo)
-    end
+    notify_rgpd_contacts_by_email(
+      responsable_traitement_email: responsable_traitement_email,
+      dpo_email: dpo_email
+    )
   end
 
   def send_application(comment:, current_user:)
     deliver_event_mailer(__method__, comment)
 
-    EnrollmentMailer.with(
-      to: enrollment.subscribers.map(&:email),
-      target_api: enrollment.target_api,
-      enrollment_id: enrollment.id,
-      template: "notify_application_sent",
-      applicant_email: current_user.email
-    ).notification_email.deliver_later
+    notify_subscribers_by_email_for_sent_application(current_user: current_user)
   end
 
   def notify(comment:, current_user:)
@@ -58,30 +46,5 @@ class BaseNotifier < AbstractNotifier
     if enrollment.dpo.present?
       deliver_rgpd_email_for(:dpo)
     end
-  end
-
-  private
-
-  def deliver_event_mailer(event, comment)
-    EnrollmentMailer.with(
-      to: enrollment.user.email,
-      target_api: enrollment.target_api,
-      enrollment_id: enrollment.id,
-      template: event.to_s,
-      message: comment
-    ).notification_email.deliver_later
-  end
-
-  def deliver_rgpd_email_for(entity)
-    RgpdMailer.with(
-      to: enrollment.public_send(entity).email,
-      target_api: enrollment.target_api,
-      enrollment_id: enrollment.id,
-      rgpd_role: Kernel.const_get("EnrollmentsController::#{entity.upcase}_LABEL"),
-      contact_label: enrollment.public_send("#{entity}_full_name"),
-      owner_email: enrollment.user.email,
-      nom_raison_sociale: enrollment.nom_raison_sociale,
-      intitule: enrollment.intitule
-    ).rgpd_contact_email.deliver_later
   end
 end
